+++ flutter/lib/common/widgets/dialog.dart
@@ -1,0 +1,74 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+
+import '../../common.dart';
+import '../../models/platform_model.dart';
+
+void changeIdDialog() {
+  var newId = "";
+  var msg = "";
+  var isInProgress = false;
+  TextEditingController controller = TextEditingController();
+  gFFI.dialogManager.show((setState, close) {
+    submit() async {
+      debugPrint("onSubmit");
+      newId = controller.text.trim();
+      setState(() {
+        msg = "";
+        isInProgress = true;
+        bind.mainChangeId(newId: newId);
+      });
+
+      var status = await bind.mainGetAsyncStatus();
+      while (status == " ") {
+        await Future.delayed(const Duration(milliseconds: 100));
+        status = await bind.mainGetAsyncStatus();
+      }
+      if (status.isEmpty) {
+        // ok
+        close();
+        return;
+      }
+      setState(() {
+        isInProgress = false;
+        msg = translate(status);
+      });
+    }
+
+    return CustomAlertDialog(
+      title: Text(translate("Change ID")),
+      content: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text(translate("id_change_tip")),
+          const SizedBox(
+            height: 12.0,
+          ),
+          TextField(
+            decoration: InputDecoration(
+                border: const OutlineInputBorder(),
+                errorText: msg.isEmpty ? null : translate(msg)),
+            inputFormatters: [
+              LengthLimitingTextInputFormatter(16),
+              // FilteringTextInputFormatter(RegExp(r"[a-zA-z][a-zA-z0-9\_]*"), allow: true)
+            ],
+            maxLength: 16,
+            controller: controller,
+            focusNode: FocusNode()..requestFocus(),
+          ),
+          const SizedBox(
+            height: 4.0,
+          ),
+          Offstage(
+              offstage: !isInProgress, child: const LinearProgressIndicator())
+        ],
+      ),
+      actions: [
+        TextButton(onPressed: close, child: Text(translate("Cancel"))),
+        TextButton(onPressed: submit, child: Text(translate("OK"))),
+      ],
+      onSubmit: submit,
+      onCancel: close,
+    );
+  });
+}
